Extract RestAuthenticator
