+++ jmix-restds/restds/src/main/java/io/jmix/restds/auth/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Haulmont.
+ * Copyright 2020 Haulmont.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package io.jmix.restds.impl;
+@Experimental
+@NonNullApi
+package io.jmix.restds.auth;
 
-public record RestConnectionParams(String baseUrl, String clientId, String clientSecret) {
-}
+import io.jmix.core.annotation.Experimental;
+import org.springframework.lang.NonNullApi;
